backend/requirements.txt
Flask_Migrate==2.7.0
Werkzeug==0.16.0
SQLAlchemy==1.3.23
Flask==1.0.2
alembic==1.5.6
Flask_SQLAlchemy==2.4.4
flask_marshmallow==0.14.0
PyJWT==2.0.1
